closes #10
